fix a split bug. avoid separate zreo by retry split. code tidy up.

+++ src/ApvlvParams.cpp
@@ -1,29 +1,29 @@
 /****************************************************************************
- * Copyright (c) 1998-2005,2006 Free Software Foundation, Inc.              
- *                                                                          
- * Permission is hereby granted, free of charge, to any person obtaining a  
- * copy of this software and associated documentation files (the            
- * "Software"), to deal in the Software without restriction, including      
- * without limitation the rights to use, copy, modify, merge, publish,      
- * distribute, distribute with modifications, sublicense, and/or sell       
- * copies of the Software, and to permit persons to whom the Software is    
- * furnished to do so, subject to the following conditions:                 
- *                                                                          
- * The above copyright notice and this permission notice shall be included  
- * in all copies or substantial portions of the Software.                   
- *                                                                          
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS  
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF               
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.   
- * IN NO EVENT SHALL THE ABOVE COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,   
- * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR    
- * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR    
- * THE USE OR OTHER DEALINGS IN THE SOFTWARE.                               
- *                                                                          
- * Except as contained in this notice, the name(s) of the above copyright   
- * holders shall not be used in advertising or otherwise to promote the     
- * sale, use or other dealings in this Software without prior written       
- * authorization.                                                           
+ * Copyright (c) 1998-2005,2006 Free Software Foundation, Inc.
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a
+ * copy of this software and associated documentation files (the
+ * "Software"), to deal in the Software without restriction, including
+ * without limitation the rights to use, copy, modify, merge, publish,
+ * distribute, distribute with modifications, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included
+ * in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
+ * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+ * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+ * IN NO EVENT SHALL THE ABOVE COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR
+ * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR
+ * THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ *
+ * Except as contained in this notice, the name(s) of the above copyright
+ * holders shall not be used in advertising or otherwise to promote the
+ * sale, use or other dealings in this Software without prior written
+ * authorization.
 ****************************************************************************/
 
 /****************************************************************************
@@ -90,7 +90,7 @@ namespace apvlv
             {
               is >> argu;
               size_t off = argu.find ('=');
-              if (off == string::npos) 
+              if (off == string::npos)
                 {
                   is >> crap >> data;
                   if (crap == "=" && data.length () > 0)
@@ -179,7 +179,7 @@ namespace apvlv
         return NULL;
       }
 
-  returnType 
+  returnType
     ApvlvParams::getmap (const char *s, int n)
       {
         map <string, string>::iterator it;
@@ -187,9 +187,9 @@ namespace apvlv
           {
             if (strncmp (it->first.c_str (), s, n) == 0)
               {
-                if (it->first.size () == n)
+                if (it->first.size () == (unsigned int) n)
                   return MATCH;
-                else 
+                else
                   return NEED_MORE;
               }
           }
